frontend/requirements.txt
Django==1.10
django-auth-ldap==1.2.8
djangorestframework==3.5.1
